Working on authorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 
 class ListingController {
@@ -15,7 +17,7 @@ class ListingController {
     }
 
     public function index() {
-        $listings = $this->db->query("SELECT * FROM listings")->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
 
         loadView("listings/index", [
@@ -64,7 +66,7 @@ class ListingController {
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map("sanitize", $newListingData);
 
@@ -110,6 +112,10 @@ class ListingController {
 
             $this->db->query($query, $newListingData);
 
+            // Set flash message
+            Session::setFlashMessage('success_message', 'Listing deleted successfully');
+
+
             redirect('/listings');
         }
     }
@@ -135,10 +141,16 @@ class ListingController {
             return;
         }
 
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authoirzed to delete this listing');
+            return redirect('/listings/' . $listing->id);
+        }
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
-        // set flash message
-        $_SESSION["success_message"] = "Listing deleted successfully";
+        // Set flash message
+        Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
         redirect('/listings');
     }
@@ -231,7 +243,7 @@ class ListingController {
             $this->db->query($updateQuery, $updateValues);
 
             // Set flash message
-            $_SESSION["success_message"] = "Listing updated";
+            Session::setFlashMessage('success_message', 'Listing updated');
 
             redirect('/listings/' . $id);
         }
